Adds typescript configuration & type definitions
I have tried to keep changes to the actual code to a minimum, and
focused on adding jsdoc type hints as much as possible. However, there
are a few places where I have made small changes - mostly to add null
checks, which typescript is particular about.

This commit trys to make no changes to the behavior of the library in
away way, beyong adding support for typescript